# impala-bridge/openapi.yaml
openapi: 3.1.0
info:
  title: impala-bridge
  version: "0.0.0"
  summary: REST bridge between offline Payala payments and the Stellar network.
  description: |
    impala-bridge issues JWTs, persists dual-chain transactions, dispatches
    notifications via SMS/email/push/webhook, and streams events from Stellar
    Horizon and Payala.

    Authentication uses a two-token strategy:

    - **Refresh token** (14-day TTL) — obtained via username+password from `POST /token`.
      Used only to mint temporal tokens, never for API access.
    - **Temporal token** (1-hour TTL) — obtained from a refresh token via `POST /token`.
      Sent as `Authorization: Bearer <token>` on every authenticated request.

    All mutating endpoints enforce account ownership via `require_owner`. A
    `POST /logout` revokes the current token (the JTI is added to a Redis
    blacklist until it would have expired naturally).

    Rate limits apply to `/authenticate`, `/token`, `/mfa/verify`, `/transaction`,
    and `/card`: 10 requests per 60-second window, keyed by account id (and
    MFA type where applicable). Exceeding the limit returns HTTP 429 with a
    `Retry-After` header.

    Global request timeout is 30 s (HTTP 408 on exceed). Postgres has a
    per-connection `statement_timeout` of 60 s.
  contact:
    name: impala-bridge maintainers
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
...
  - url: https://{host}
...
security:
  - bearerAuth: []
paths:
...
  /auth/okta:
    post:
      tags: [auth]
      summary: Exchange an Okta access token for bridge tokens (legacy single-provider flow).
      description: |
        Validates the Okta access token, auto-provisions the account on first
        login, and issues a refresh + temporal token pair. With
        `cookie_mode: true` (used by the web UI) it instead establishes an
        HttpOnly cookie session and returns a `SessionResponse`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OktaTokenExchangeRequest' }
      responses:
        '200':
          description: Bridge tokens (bearer mode) or session identity (cookie mode).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TokenResponse'
                  - $ref: '#/components/schemas/SessionResponse'
        '400': { description: Okta authentication is not configured. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Temporal token (1-hour TTL). Obtain via `POST /token` with a valid
        refresh token. Send as `Authorization: Bearer <token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: impala_session
      description: |
        Browser session cookie (HttpOnly), established by `POST /session/login`
        or `POST /auth/okta` with `cookie_mode: true`. In production the cookie
        is `__Host-`-prefixed (`__Host-impala_session`). Unsafe-method requests
        additionally require the CSRF synchronizer token in the `X-CSRF-Token`
        header.
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "unauthorized" }
            message: { type: string }
          required: [code, message]
      required: [error]
...
    TokenResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        refresh_token: { type: string, nullable: true }
        temporal_token: { type: string, nullable: true }
...
    SessionResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        account_id: { type: string }
        is_admin: { type: boolean }
        csrf_token:
          type: string
          description: CSRF synchronizer token for `X-CSRF-Token` on unsafe-method cookie requests.
    OktaTokenExchangeRequest:
      type: object
      properties:
        okta_token: { type: string, description: Okta access token. }
        cookie_mode:
          type: boolean
          default: false
          description: Establish an HttpOnly cookie session instead of returning bearer tokens.
      required: [okta_token]
...